feat: make snake movement
The snake was static and part of the background.

This commit make the movement of the snake possible. it implement a
queue and to move, it remove the tail and place it on the front with
its position updated based on the snake's direction.

To make the control and drawing of the snake, the snake was made a
global variable.

Also make snake growth possible.

Fixes #18